✨[#172] Feat: 대시보드 컬럼 에러바운더리 처리
