+++ features/dashboard/dashboard-list/components/index.ts
@@ -2,6 +2,5 @@ import DashboardList from "./list/dashboard-list"
 import DashboardListItem from "./list-item/dashboard-list-item"
 import DashboardListPagination from "./pagination/dashboard-list-pagination"
 import DashboardListSkeleton from "./list-skeleton/dashboard-list-skeleton"
-import DashboardErrorFallback from "./error/dashboard-error-fallback"
 
-export { DashboardList, DashboardListItem, DashboardListPagination, DashboardListSkeleton, DashboardErrorFallback }
+export { DashboardList, DashboardListItem, DashboardListPagination, DashboardListSkeleton }
