+++ pages/dashboard/my/index.tsx
@@ -1,12 +1,9 @@
 import { Meta } from "@common/components"
-import { DashboardLayout, DashboardHeader, DashboardSideBar, ErrorBoundary } from "@shared/dashboard/components"
-import {
-  DashboardErrorFallback,
-  DashboardList,
-  DashboardListPagination,
-} from "@features/dashboard/dashboard-list/components"
+import { DashboardLayout, DashboardHeader, DashboardSideBar } from "@shared/dashboard/components"
+import { DashboardList, DashboardListPagination } from "@features/dashboard/dashboard-list/components"
 import { DashboardInvited } from "@features/dashboard/dashboard-invited/components"
 import { useDashboardPageStore, useFetchDashboards } from "@shared/dashboard/hooks"
+import { DashboardErrorFallback, ErrorBoundary } from "@shared/dashboard/components"
 
 export default function MyDashboard() {
   const currentPage = useDashboardPageStore.use.currentPage()
