+++ shared/dashboard/components/dashboard-layout/dashboard/dashboard-layout.module.css
@@ -5,6 +5,7 @@
   padding-left: 40px;
   padding-bottom: 60px;
   overflow-x: scroll;
+  position: relative;
 }
 
 @media (max-width: 1400px) {
