+++ shared/dashboard/components/error-boundary/dashboard-error-boundary.tsx
@@ -10,7 +10,7 @@ interface ErrorBoundaryState {
   hasError: boolean
 }
 
-export class ErrorBoundary extends React.Component<ErrorBoundaryProps, ErrorBoundaryState> {
+export default class ErrorBoundary extends React.Component<ErrorBoundaryProps, ErrorBoundaryState> {
   constructor(props: ErrorBoundaryProps) {
     super(props)
     this.state = { hasError: false }
