+++ shared/dashboard/components/error-boundary/index.ts
@@ -1,0 +1,4 @@
+import ErrorBoundary from "./dashboard-error-boundary"
+import DashboardErrorFallback from "./dashboard-error-fallback"
+
+export { ErrorBoundary, DashboardErrorFallback }
