+++ shared/dashboard/components/index.ts
@@ -1,4 +1,4 @@
 export * from "./dashboard-header"
 export * from "./dashboard-layout"
 export * from "./dashboard-sidebar"
-export * from "./error-boundary/dashboard-error-boundary"
+export * from "./error-boundary"
